esp32: shift to support latest esp-idf
Esp-idf has an improved but still kind of abused cmake-
based build system now.

If we see ESP_PLATFORM coming as a cmake var, we can know we
are being built from inside the esp-idf config system.

Leave the existing esp32 arrangements alone but triggered off
ESP_PLATFORM, adapt to use the cross toolchain file and
various quirks automatically.

In this way you can build lws a part of your project in a
much cleaner way.

Prepare a minimal esp32 test app for use in Sai

Adapt .sai.json to build for esp32

+++ lib/plat/freertos/freertos-file.c
@@ -135,7 +135,7 @@ lws_find_string_in_file(const char *filename, const char *string, int stringlen)
 
 	q = string;
 	n = 0;
-	while (n < sizeof(buf) - 1 && q != p)
+	while ((size_t)n < sizeof(buf) - 1 && q != p)
 		buf[n++] = *q++;
 	buf[n] = '\0';
